- fix DECLARE_(FUN|PROC)[789] - top_xxx -> xxx_count - increment pObjIndex->reset_num when resetting objects - use varr_size - unsigned char -> u_char - unsigned -> size_t where appropriate - everything can now be compiled with BDECFLAGS, turn on -Werror by default
git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@3004 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/mod_bootdb/db_lang.c
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: db_lang.c,v 1.32 2001-08-05 16:36:28 fjoe Exp $
+ * $Id: db_lang.c,v 1.33 2001-08-14 16:06:48 fjoe Exp $
  */
 
 #include <stdio.h>
@@ -163,7 +163,7 @@ DBLOAD_FUN(load_rulecl)
 				if (s) {
 					strnzncpy(path, sizeof(path),
 					    bootdb_filename,
-					    (unsigned)(s - bootdb_filename));
+					    (size_t) (s - bootdb_filename));
 				} else
 					path[0] = '\0';
